+++ .babelrc
@@ -1,6 +1,6 @@
 {
   "presets": [
-    "babel-preset-react-native-stage-0/decorator-support"
+    "babel-preset-react-native-stage-0/decorator-support", "babel-preset-expo"
   ],
   "env": {
     "development": {
