+++ componets/Main.js
@@ -1,0 +1,3 @@
+import Main from './app/index';
+
+export default Main;
