+++ componets/app/components/Card/index.js
@@ -1,0 +1,4 @@
+import Card from './Card';
+import styles from './styles';
+
+export { Card, styles };
