+++ componets/app/screens/ModalPickerImage/style.js
